// apps/api/hasura-graphql.ts
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  timestamptz: any;
  uuid: any;
};
// ...
export type CreateUserMutationVariables = {
  email: Scalars['String'];
  password: Scalars['String'];
};